Feat(WinningNumber): Add WinningNumber domain and move validation logic inside
