+++ src/main/java/lotto/Validator.java
@@ -6,28 +6,6 @@ import java.util.Set;
 
 public class Validator {
 
-    public void validateWinningNumbersCount(List<Integer> winningNumbers) {
-        if (winningNumbers.size() != 6) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public void validateWinningNumbersRange(List<Integer> winningNumbers) {
-        for (Integer winningNumber : winningNumbers) {
-            if (winningNumber > 45 || winningNumber < 1) {
-                throw new IllegalArgumentException();
-            }
-        }
-    }
-
-    public void validateWinningNumbersDuplication(List<Integer> winningNumbers) {
-        int originalSize = winningNumbers.size();
-        Set<Integer> winningNumbersSet = new HashSet<>(winningNumbers);
-        if (originalSize != winningNumbersSet.size()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
     public void validateBonusNumberRange(int bonusNumber) {
         if (bonusNumber > 45 || bonusNumber < 1) {
             throw new IllegalArgumentException();
